refactor(db): skill ut ferdigstilling av oppfølgingsplan
Flytt påminnelsespolicy og metrikker ut av repository-laget uten å dele transaksjonen.
