+++ src/test/kotlin/no/nav/syfo/dokumentporten/DokumentportenServiceTest.kt
@@ -11,7 +11,7 @@ import no.nav.syfo.TestDB
 import no.nav.syfo.aareg.AaregService
 import no.nav.syfo.defaultPersistedOppfolgingsplan
 import no.nav.syfo.dokumentporten.client.FakeDokumentportenClient
-import no.nav.syfo.oppfolgingsplan.db.OppfolgingsplanEvalueringPaaminnelseRepository
+import no.nav.syfo.oppfolgingsplan.db.OppfolgingsplanFinalizationRepository
 import no.nav.syfo.oppfolgingsplan.db.domain.PersistedOppfolgingsplan
 import no.nav.syfo.oppfolgingsplan.db.findOppfolgingsplanerForDokumentportenPublisering
 import no.nav.syfo.oppfolgingsplan.db.setNarmesteLederFullName
@@ -33,7 +33,7 @@ class DokumentportenServiceTest :
             esyfovarselProducer = mockk<EsyfovarselProducer>(relaxed = true),
             aaregService = mockk<AaregService>(relaxed = true),
             unntaksvurderingService = mockk(relaxed = true),
-            oppfolgingsplanEvalueringPaaminnelseRepository = OppfolgingsplanEvalueringPaaminnelseRepository(testDb),
+            oppfolgingsplanFinalizationRepository = OppfolgingsplanFinalizationRepository(testDb),
         )
 
         beforeTest {
