+++ src/test/kotlin/no/nav/syfo/oppfolgingsplan/api/v1/sykmeldt/OppfolgingsplanApiV1Test.kt
@@ -42,7 +42,7 @@ import no.nav.syfo.isdialogmelding.client.IsDialogmeldingClient
 import no.nav.syfo.istilgangskontroll.IsTilgangskontrollService
 import no.nav.syfo.istilgangskontroll.client.IIsTilgangskontrollClient
 import no.nav.syfo.oppfolgingsplan.api.v1.registerApiV1
-import no.nav.syfo.oppfolgingsplan.db.OppfolgingsplanEvalueringPaaminnelseRepository
+import no.nav.syfo.oppfolgingsplan.db.OppfolgingsplanFinalizationRepository
 import no.nav.syfo.oppfolgingsplan.db.persistUnntaksvurdering
 import no.nav.syfo.oppfolgingsplan.db.softDeleteExpiredUnntaksvurderinger
 import no.nav.syfo.oppfolgingsplan.dto.FerdigstiltPlanHendelse
@@ -119,8 +119,8 @@ class OppfolgingsplanApiV1Test :
                                 pdlService = pdlServiceMock,
                                 aaregService = aaregServiceMock,
                                 unntaksvurderingService = unntaksvurderingService,
-                                oppfolgingsplanEvalueringPaaminnelseRepository =
-                                OppfolgingsplanEvalueringPaaminnelseRepository(testDb),
+                                oppfolgingsplanFinalizationRepository =
+                                OppfolgingsplanFinalizationRepository(testDb),
                             ),
                             unntaksvurderingService = unntaksvurderingService,
                             pdfGenService = pdfGenService,
